Query VFB's Neo4J graph database

Query VFB's Neo4J graph database

Usage

vfb_neo4j_query(
  x,
  ...,
  path = "db/data",
  server = getOption("vfbr.server.neo4j")
)

Arguments

x

A character query in Neo4J's cypher language

...

Additional query arguments of the form key=value

path

The relative path on the server for the Neo4J endpoint

server

The server's root URL

Value

A data.frame of query results

Details

Under the hood, this uses the RNeo4j::cypher function to call a Neo4J service running on the specified VFB server.

References

https://neo4j.com/developer/cypher-query-language/

See also

Examples

# \donttest{
# ask for all neuronal classes
nclasses=vfb_neo4j_query("MATCH (n:Neuron:Class) RETURN n.label")
#> Warning: changing locked binding for ‘length.path’ in ‘httr’ whilst loading ‘RNeo4j’
#> Error in vfb_neo4j_query("MATCH (n:Neuron:Class) RETURN n.label"): Unable to connect to VFB neo4j server: http://pdb.virtualflybrain.org
nrow(nclasses)
#> Error in eval(expr, envir, enclos): object 'nclasses' not found
head(nclasses)
#> Error in eval(expr, envir, enclos): object 'nclasses' not found

# Find all images with an associated neuronal class
q=paste0("MATCH (n:Class:VFB { label : 'neuron' })",
  "<-[:SUBCLASSOF*]-(p)<-[:INSTANCEOF]-(i:Individual)",
  "RETURN distinct i.label, p.label;")
nclasses_image=vfb_neo4j_query(q)
#> Error in vfb_neo4j_query(q): Unable to connect to VFB neo4j server: http://pdb.virtualflybrain.org
nrow(nclasses_image)
#> Error in eval(expr, envir, enclos): object 'nclasses_image' not found
head(nclasses_image)
#> Error in eval(expr, envir, enclos): object 'nclasses_image' not found

# how many neuronal classes have images?
length(unique(nclasses_image$p.label))
#> Error in eval(expr, envir, enclos): object 'nclasses_image' not found

# look at number of images for each neuronal class
table_by_nclass <- table(nclasses_image$p.label)
#> Error in eval(expr, envir, enclos): object 'nclasses_image' not found
subset(as.data.frame(table_by_nclass), Freq>200)
#> Error in eval(expr, envir, enclos): object 'table_by_nclass' not found
# }