Find canonical ontology term using VFB SOLR to query synonyms

Find canonical ontology term using VFB SOLR to query synonyms

Usage

vfb_synonym_query(
  query,
  exact = TRUE,
  quote = NA,
  searchfields = c("synonym", "label"),
  fields = "short_form label synonym",
  verbose = interactive(),
  ...
)

Arguments

query

A character vector containing one or more queries - see details. Wildcards (?*) can be used, and have a similar meaning to their use in the shell. Queries are case insensitive.

exact

Whether to do an exact match (i.e. the query must match the whole string, the default) or to allow partial matches against synonyms. See Query details and examples.

quote

Whether to quote the query so that individual terms must all be matched in the order given or to OR each term in the query. The default quote=NA will quote when exact=TRUE and the query does not contain wildcards. See Query details and Examples.

searchfields

Character vector specifying fields to search. The default searches both synonyms and the canonical term label (since the canonical term is not included in the synonym list).

fields

The fields to return (defaults to 'short_form label synonym'). Setting fields="" implies all fields (but note that not all results may have the same fields - see discussion in details)

verbose

Whether to print messages to the screen.

...

Additional arguments passed to pbsapply when there are multiple input queries and eventually vfb_solr_query. You can use this e.g. to set the number of returned rows.

Value

A data.frame containing one or more result rows, ordered according to the solr result score, with attributes

Details

Note that data.frame that is returned may contain a list in the synonym column because there will likely be multiple synonyms for a given query.

When x has length > 1 i.e. multiple query terms, then multiple calls to vfb_solr_query are wrapped in a pbsapply statement. You can pass arguments to sapply in ... such as simplify=FALSE if you wish. By default the return value will be a matrix with rows that you can index by field name and columns that you can index by queries. However the form of this is not guaranteed if you ask for fields that are present only for some of the results.

Note also that the

Query details

The exact and quote arguments have a profound effect on how SOLR carries out searches. Setting both arguments to FALSE gives the least specific search and you may want to do this if you do not find any results with an initial query using the default values.

In general SOLR will break any text in fields like synonym into separate tokens based on whitespace, hyphens etc, so the value :

inter-antennal lobe tract

will map onto 4 tokens.

inter antennal lobe tract

This means that a search with query "inter antennal lobe tract", exact=F will be successful (note missing hyphen).

When exact=TRUE, the values will not be tokenised and the query must match the whole field.

The quote argument modifies the query. By default the query is tokenised as just described for fields in the database, tokens can be matched in any order and only one token must match (i.e. the are ORed together). When quote=TRUE, the query must match in the order given (this is achieved by wrapping it in double quotes before passing to SOLR). For example

vfb_synonym_query("antennal lobe tract", exact=F, quote=T)

returns 2 results at the time of writing whereas

vfb_synonym_query("antennal lobe tract", exact=F, quote=F)

returns 194 results.

See https://github.com/EBISPOT/OLS/blob/master/ols-solr/src/main/solr-conf/ontology/conf/schema.xml for definition of the of the synonym_s exact match field.

See also

Examples

vfb_synonym_query("SOG")
#> data frame with 0 columns and 0 rows
# anything with a synonym that includes SOG
vfb_synonym_query("SOG", exact=FALSE)
#>                                                        label    short_form
#> 1              adult dopaminergic neuron of gnathal ganglion FBbt_00048266
#> 2  male specific 1 neuron of subesophageal ganglion (female) FBbt_00048031
#> 3    male specific 1 neuron of subesophageal ganglion (male) FBbt_00048030
#> 4                                    gnathal ganglion neuron FBbt_00052032
#> 5                                                        sog   FBgn0003463
#> 6                                   adult subesophageal zone FBbt_00110639
#> 7                          adult antenno-subesophageal tract FBbt_00007354
#> 8                           larval subesophageal CCAP neuron FBbt_00111326
#> 9                      larval CCAP subesophageal IN-1 neuron FBbt_00111345
#> 10                                          gnathal ganglion FBbt_00004013
#> 11                                 subesophageal zone neuron FBbt_00001972
#> 12                                octopaminergic VUMa neuron FBbt_00110155
#> 13     adult dopaminergic VUM neuron of the gnathal ganglion FBbt_00110128
#> 14                                    adult gnathal ganglion FBbt_00014013
#> 15                                       larval NPF-SOG cell FBbt_00047253
#> 16                           larval subesophageal VUM neuron FBbt_00047750
#> 17                  cell body rind of adult gnathal ganglion FBbt_00100210
#> 18                                   larval gnathal ganglion FBbt_00100138
#> 19               obsolete SOG ventral unpaired median neuron FBbt_00100785
#>                                                                                                                                                                                                                                                                                                                              synonym
#> 1                                                                                                                                                                                          adult dopaminergic neuron of gnathal ganglion, adult dopaminergic SB neuron, adult dopaminergic GNG neuron, adult dopaminergic SOG neuron
#> 2                                                                                                                                                         female octopaminergic MS1, female octopaminergic MS1 neuron of subesophageal ganglion, female MS1 neuron of SOG, male specific 1 neuron of subesophageal ganglion (female)
#> 3                                                                                                                                                                 male octopaminergic MS1, male octopaminergic MS1 neuron of subesophageal ganglion, male specific 1 neuron of subesophageal ganglion (male), male MS1 neuron of SOG
#> 4                                                                                                                                                                                                                                          SOG neuron, subesophageal ganglion neuron, SEG neuron, GG neuron, gnathal ganglion neuron
#> 5                                                                                                                                                     Short gastrulation, l(1)G0160, short gastrulation, Sog, CG9224, short-gastrulation, l(1)G0479, Chordin, SOG, l(1)G0395, short of gastrulation, sog, Dm sog, Short Gastrulation
#> 6                                                                                                                                                                                                                               adult subesophageal zone, adult subesophageal ganglion, adult suboesophageal ganglion, SOG, SEG, SEZ
#> 7                                                                                                                                                                          antenno-suboesophageal tract, antennal lobe-subesophageal tract, adult antenno-subesophageal tract, antennal lobe-SEG tract, antennal lobe-SOG tract, AST
#> 8                                                                                                                                                                                                                                                                               IN SOG, LSE neuron, larval subesophageal CCAP neuron
#> 9                                                                                                                                                                                                                                                                                    larval CCAP subesophageal IN-1 neuron, IN SOG-1
#> 10                                                                                                                                                                                                                                                  gnathal ganglion, subesophageal ganglion, suboesophageal ganglion, GNG, SOG, SEG
#> 11                                                                                                                                                                                                                                      SOG neuron, SEG neuron, subesophageal ganglion neuron, subesophageal zone neuron, SEZ neuron
#> 12                                                             octopaminergic VUMa neuron, VUM neuron of the subesophageal ganglion, subesophageal ganglion VUM neuron, OA-VUMa, ventral unpaired median octopaminergic anterior neuron, subesophageal ganglion ventral unpaired median neuron, ascending VUM neuron, SOG VUM neuron
#> 13               TH-VUM, adult dopaminergic SB VUM neuron, adult dopaminergic VUM neuron of the gnathal ganglion, VUM neuron of the dopaminergic GNG cluster, adult subesophageal ventral unpaired median dopaminergic neuron, SOG dopaminergic neuron, dopaminergic VUM neuron, tyrosine hydroxylase ventral unpaired median neuron
#> 14                                                                                                                                                                                                                                adult subesophageal ganglion, ipsl, adult gnathal ganglion, suboesophageal ganglion, GNG, SOG, SEG
#> 15                                                                                                                                                                                                                                          larval NPF-SOG neuron, larval dNPF-SOG cell, larval NPF-SOG cell, larval dNPF-SOG neuron
#> 16                                                                                                                                                                                                 larval subesophageal VUM neuron, larval subesophageal ganglion ventral unpaired median neuron, larval SOG VUM neuron, larval sVUM
#> 17 cortex of subesophageal ganglion, rGNG, cortex of suboesophageal ganglion, SOG cortex, cell body rind around the GNG, cell body rind of adult gnathal ganglion, GCBR, cell body rind around the gnathal ganglion, suboesophageal ganglion cell body rind, cortex of adult gnathal ganglion, subesophageal ganglion cell body rind
#> 18                                                                                               embryonic/larval subesophageal ganglion, larval gnathal ganglion, larval suboesophageal ganglion, larval subesophageal zone, embryonic/larval SOG, larval GNG, larval SOG, hindbrain, larval SEG, embryonic/larval gnathal ganglion
#> 19                                                                                                                                   VUM neuron of the subesophageal ganglion, obsolete SOG ventral unpaired median neuron, subesophageal ganglion VUM neuron, subesophageal ganglion ventral unpaired median neuron, SOG VUM neuron

# query for PAM exactly
vfb_synonym_query("PAM")
#> data frame with 0 columns and 0 rows
# any term with a synonym containing PAM
vfb_synonym_query("PAM", exact = FALSE)
#>                                                label         short_form
#> 1             peptidylglycine monooxygenase activity         GO_0004504
#> 2                PAM01 neuron of upper MB commissure      FBbt_00049930
#> 3              PAM01 neuron of ventral MB commissure      FBbt_00049931
#> 4                              PAM01 feedback neuron      FBbt_00049932
#> 5                         dopaminergic PAM neuron 15      FBbt_00049841
#> 6                    larval dopaminergic pPAM neuron      FBbt_00111633
#> 7                          dopaminergic PAM neuron 9      FBbt_00111021
#> 8                          dopaminergic PAM neuron 8      FBbt_00111020
#> 9                          dopaminergic PAM neuron 6      FBbt_00111018
#> 10                         dopaminergic PAM neuron 5      FBbt_00111017
#> 11                         dopaminergic PAM neuron 2      FBbt_00111016
#> 12                         dopaminergic PAM neuron 1      FBbt_00111015
#> 13                         dopaminergic PAM neuron 7      FBbt_00111019
#> 14                        dopaminergic PAM neuron 14      FBbt_00111024
#> 15                        dopaminergic PAM neuron 13      FBbt_00111023
#> 16                        dopaminergic PAM neuron 11      FBbt_00111022
#> 17                         dopaminergic PAM neuron 4      FBbt_00110993
#> 18                        dopaminergic PAM neuron 10      FBbt_00110994
#> 19 PRESEQUENCE TRANSLOCASE-ASSOCIATED MOTOR SUBUNITS        FBgg0001908
#> 20    TH negative neuron associated with PAM cluster      FBbt_00047184
#> 21               P{GMR10G03-GAL4} expression pattern VFBexp_FBtp0057363
#> 22               P{GMR48H11-GAL4} expression pattern VFBexp_FBtp0060750
#> 23               P{GMR58E02-GAL4} expression pattern VFBexp_FBtp0061564
#> 24                        dopaminergic PAM neuron 12      FBbt_00100230
#> 25                         dopaminergic PAM neuron 3      FBbt_00100231
#> 26                           dopaminergic PAM neuron      FBbt_00100214
#> 27                    PAM-y5(uc)#5_R (FAFB:11149656)       VFB_001014mo
#> 28                    PAM-y5(uc)#4_R (FAFB:11274761)       VFB_001014mt
#> 29                    PAM-y5(da)#1_R (FAFB:11169413)       VFB_001014mp
#> 30                    PAM-y5(uc)#6_R (FAFB:11175371)       VFB_001014mq
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                       synonym
#> 1  peptidylglycine 2-hydroxylase activity, peptidylglycine alpha-hydroxylase activity, peptidylglycine monooxygenase activity, peptide-alpha-amide synthetase activity, PAM-B, peptidylglycine alpha-amidating monooxygenase activity, PAM-A, peptidyl alpha-amidating enzyme activity, peptide alpha-amidating enzyme, PAM activity, peptide alpha-amide synthase activity, synthase, peptide alpha-amide, peptidylglycine,ascorbate:oxygen oxidoreductase (2-hydroxylating)
#> 2                                                                                                                                                                                                                                                                                                 PAM-gamma5 upper commissure neuron, PAM-gamma5(uc), gamma5-uc, PAM-gamma5(da), PAM01 neuron of upper MB commissure, PAM01-uc, dopaminergic PAM neuron 1 of upper commissure
#> 3                                                                                                                                                                     PAM-gamma5(ba), PAM01 neuron of middle commissure, PAM01 neuron of ventral MB commissure, PAM01 neuron of lower commissure, PAM-gamma5 lower commissure neuron, PAM01-lc, PAM-gamma5 middle commissure neuron, gamma5-lc, dopaminergic PAM neuron 1 of ventral mushroom body commissure, PAM-gamma5(lc)
#> 4                                                                                                                                                                                                                                                                                                                                PAM01 feedback neuron, PAM01-fb, PAM-gamma5 dorsal dendrite neuron, PAM-gamma5(fb), PAM-gamma5(dd), PAM-gamma5 feedback neuron, gamma5-lc-fb
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                PAM15, dopaminergic PAM neuron 15, PAM-15, PAM-gamma5beta'2a
#> 6                                                                                                                                                                                                                                                                                                                                larval dopaminergic pPAM neuron, larval PAM neuron, larval DAN of pPAM cluster, primary-lineage protocerebral anterior medial cluster neuron
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                      PAM-beta1ped, PAM09, PAM-09, dopaminergic PAM neuron 9
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                        dopaminergic PAM neuron 8, PAM08, PAM-08, PAM-gamma4
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                       PAM-beta'2m, PAM06, PAM-06, dopaminergic PAM neuron 6
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                      dopaminergic PAM neuron 5, PAM-beta'2p, PAM-05, PAM05
#> 11                                                                                                                                                                                                                                                                                                                                                                                                         beta'2a DAN, dopaminergic PAM neuron 2, PAM-beta'2a, PAM02, PAM-02
#> 12                                                                                                                                                                                                                                                                                                                                                                              PAM-01, adult fruitless aSP13 neuron, PAM01, DAN-aSP13, dopaminergic PAM neuron 1, PAM-gamma5
#> 13                                                                                                                                                                                                                                                                                                                                                                                                          PAM-07, PAM-gamma4-gamma1gamma2, PAM07, dopaminergic PAM neuron 7
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                     PAM-14, dopaminergic PAM neuron 14, PAM-beta'1m, PAM14
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                    PAM-beta'1ap, PAM13, dopaminergic PAM neuron 13, PAM-13
#> 16                                                                                                                                                                                                                                                                                                                                                                                                              dopaminergic PAM neuron 11, MB-VP1, PAM11, PAM-11, PAM-alpha1
#> 17                                                                                                                                                                                                                                                                                                                                                                                                                        dopaminergic PAM neuron 4, PAM-04, PAM-beta2, PAM04
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                       dopaminergic PAM neuron 10, PAM-10, PAM-beta1, PAM10
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                     PRESEQUENCE TRANSLOCASE-ASSOCIATED MOTOR SUBUNITS, PAM
#> 20                                                                                                                                                                                                                                                                                                                                                                                                                             TH negative neuron associated with PAM cluster
#> 21                                                                                                                                                                                                                                                                                                 P{GMR10G03-GAL4} expression pattern, PAM-DAN-γ4-Gal4 expression pattern, R10G03-Gal4 expression pattern, PAM-DAN-gamma4-Gal4 expression pattern, R10G03 expression pattern
#> 22                                                                                                                                                                                                                                                                PAM-DAN-γ5-Gal4 expression pattern, PAM-DAN-gamma5-Gal4 expression pattern, R48H11-Gal4 expression pattern, R48H11 expression pattern, P{GMR48H11-GAL4} expression pattern, R48H011-Gal4 expression pattern
#> 23                                                                                                                                                                          R58E02GAL4 expression pattern, R58E02-Gal4 expression pattern, GMR58E02-Gal4 expression pattern, GMR58E02-GAL4 expression pattern, R58E02 expression pattern, R58E02-GAL4 expression pattern, P{GMR58E02-GAL4} expression pattern, 58E02-GAL4 expression pattern, PAM-DAN-Gal4 expression pattern
#> 24                                                                                                                                                                                                                                                                                                                                                         mushroom body medial lobe arborizing neuron 2, MB-M2 neuron, PAM-12, PAM12, dopaminergic PAM neuron 12, PAM-gamma3
#> 25                                                                                                                                                                                                                                                                                                                                                           MB-M3, PAM-03, PAM-beta2beta'2a, dopaminergic PAM neuron 3, PAM03, mushroom body medial lobe arborizing neuron 3
#> 26                                                                                                                                                                                                                                     protocerebral anteriomedial dopaminergic cluster neuron, PAM, AM cell cluster neuron, dopaminergic PAM neuron, TH-positive anteriomedial cluster neuron, PAM cluster neuron, protocerebral anterior medial dopaminergic cluster neuron
#> 27                                                                                                                                                                                                                                                                                                                                                                                                                                      Putative DAN PAM-y5 (25) 11149657 KJH
#> 28                                                                                                                                                                                                                                                                                                                                                                                                                                      Putative DAN PAM-y5 (24) 11274762 KJH
#> 29                                                                                                                                                                                                                                                                                                                                                                                                                                      Putative DAN PAM-y5 (21) 11169414 KJH
#> 30                                                                                                                                                                                                                                                                                                                                                                                                                                     Putative DAN PAM-y5 (22) 11175372 AJES

# terms with synonyms that start with "PAM-"
vfb_synonym_query("PAM-*")
#> data frame with 0 columns and 0 rows
# nb this doesn't return any results when exact=FALSE because solr matches
# against tokenised strings (i.e. strings that have broken on spaces and
# other non word characters such as dashes and underscores).
vfb_synonym_query("PAM-*", exact = FALSE)
#>                              label
#> 1   PAM-y5(uc)#5_R (FAFB:11149656)
#> 2   PAM-y5(uc)#4_R (FAFB:11274761)
#> 3   PAM-y5(da)#1_R (FAFB:11169413)
#> 4   PAM-y5(uc)#6_R (FAFB:11175371)
#> 5   PAM-y5(uc)#3_R (FAFB:11262395)
#> 6   PAM-b'2(c1)#1_R (FAFB:8317886)
#> 7  PAM-b'2a(c2)#1_R (FAFB:7777207)
#> 8      PAM-b'2a#1_R (FAFB:6969526)
#> 9  PAM-b'2a(nc)#1_R (FAFB:7270967)
#> 10 PAM-b'2a(c2)#2_R (FAFB:7512499)
#> 11   PAM-y5(lc)#4_R (FAFB:6019056)
#> 12   PAM-y5(ba)#2_R (FAFB:6059582)
#> 13       PAM-y5#9_L (FAFB:6100678)
#> 14       PAM-y5#2_L (FAFB:6109747)
#> 15     PAM-b'2a#4_R (FAFB:6116676)
#> 16   PAM-y5(uc)#1_R (FAFB:6181246)
#> 17     PAM-b'2a#2_R (FAFB:5886218)
#> 18       PAM-y5#5_L (FAFB:5650098)
#> 19   PAM-y5(lc)#1_R (FAFB:5652208)
#> 20   PAM-y5(lc)#8_R (FAFB:5631471)
#> 21     PAM-b'2a#3_R (FAFB:5737978)
#> 22       PAM-y5#6_L (FAFB:5737998)
#> 23       PAM-y5#8_L (FAFB:6013565)
#> 24       PAM-y5#7_L (FAFB:5738037)
#> 25   PAM-y5(ba)#1_R (FAFB:6013704)
#> 26   PAM-y5(lc)#2_R (FAFB:5914197)
#> 27   PAM-y5(lc)#3_R (FAFB:5952626)
#> 28       PAM-y5#3_L (FAFB:5470844)
#> 29   PAM-y5(fb)#3_R (FAFB:5234721)
#> 30       PAM-y5#1_L (FAFB:5262509)
#>                                                          synonym   short_form
#> 1                          Putative DAN PAM-y5 (25) 11149657 KJH VFB_001014mo
#> 2                          Putative DAN PAM-y5 (24) 11274762 KJH VFB_001014mt
#> 3                          Putative DAN PAM-y5 (21) 11169414 KJH VFB_001014mp
#> 4                         Putative DAN PAM-y5 (22) 11175372 AJES VFB_001014mq
#> 5                          Putative DAN PAM-y5 (23) 11262396 KJH VFB_001014ms
#> 6       Putative DAN PAM-B'2a (4) 1683766 Frog Cake MWP AJES KJH VFB_001014l8
#> 7            Putative DAN PAM-B'2a (1) 5712028 Anzac Biscuit KJH VFB_001014kh
#> 8         Putative DAN PAM-B'2a neuron 6969527 Sweet Creature RT VFB_001014ig
#> 9  Putative DAN PAM-B'2a (2) 7270968 Chocolate Crackles AJES KJH VFB_001014k1
#> 10             Putative DAN PAM-B'2a (3) 7512500 Fairy Bread KJH VFB_001014k9
#> 11                    Putative PAM-y5 6019057 MWP (16) Olive MWP VFB_001014ga
#> 12                        Putative DAN PAM-y5 (17) 6059583 AJES  VFB_001014gg
#> 13                fragment 6100679 AJES putative DAN PAM-y5 LEFT VFB_001014gn
#> 14           putative DAN PAM-y5 (26) 6109748 MWP Sojourner LEFT VFB_001014go
#> 15                             Putative DAN PAM 6116677 KJH AJES VFB_001014gv
#> 16                                  DAN PAM-y5 (19) 6181247 AJES VFB_001014hf
#> 17                                   DAN PAM-y5 5886219 JH RIGHT VFB_001014f8
#> 18                      putative DAN PAM-y5 (9) 5650099 KJH LEFT VFB_001014el
#> 19                           putative DAN PAM-y5 (10) 5652209 NM VFB_001014em
#> 20                           putative DAN PAM-y5 (8) 5631472 KJH VFB_001014ej
#> 21               fragment 5737979 AJES Putative DAN/MBON B'2/B2? VFB_001014et
#> 22                    Putative DAN PAM-y5 (14) 5737999 AJES LEFT VFB_001014eu
#> 23               Putative DAN PAM-y5 (18) 6013566 Joyce MWP LEFT VFB_001014g7
#> 24          large fragment 5738038 AJES putative DAN PAM-y5 LEFT VFB_001014ev
#> 25                   Putative DAN PAM-y5 (15) 6013705 AJES RIGHT VFB_001014g8
#> 26                  Putative DAN PAM-y5 (12) 5914197 Dead End NO VFB_001014fb
#> 27                         Putative DAN PAM-y5 (13) 5952627 AJES VFB_001014fn
#> 28                     putative DAN PAM-y5 (11) 5470845 KJH LEFT VFB_001014e8
#> 29          Putative DAN PAM-y5 (6) 5234722 Capitol Building CMM VFB_001014dd
#> 30                fragment 5262510 AJES putative DAN PAM-y5 LEFT VFB_001014dk
# However if you quote the query you will get results
vfb_synonym_query("PAM-*", exact = FALSE, quote=TRUE)
#>                                                label         short_form
#> 1             peptidylglycine monooxygenase activity         GO_0004504
#> 2                PAM01 neuron of upper MB commissure      FBbt_00049930
#> 3              PAM01 neuron of ventral MB commissure      FBbt_00049931
#> 4                              PAM01 feedback neuron      FBbt_00049932
#> 5                         dopaminergic PAM neuron 15      FBbt_00049841
#> 6                    larval dopaminergic pPAM neuron      FBbt_00111633
#> 7                          dopaminergic PAM neuron 9      FBbt_00111021
#> 8                          dopaminergic PAM neuron 8      FBbt_00111020
#> 9                          dopaminergic PAM neuron 6      FBbt_00111018
#> 10                         dopaminergic PAM neuron 5      FBbt_00111017
#> 11                         dopaminergic PAM neuron 2      FBbt_00111016
#> 12                         dopaminergic PAM neuron 1      FBbt_00111015
#> 13                         dopaminergic PAM neuron 7      FBbt_00111019
#> 14                        dopaminergic PAM neuron 14      FBbt_00111024
#> 15                        dopaminergic PAM neuron 13      FBbt_00111023
#> 16                        dopaminergic PAM neuron 11      FBbt_00111022
#> 17                         dopaminergic PAM neuron 4      FBbt_00110993
#> 18                        dopaminergic PAM neuron 10      FBbt_00110994
#> 19 PRESEQUENCE TRANSLOCASE-ASSOCIATED MOTOR SUBUNITS        FBgg0001908
#> 20    TH negative neuron associated with PAM cluster      FBbt_00047184
#> 21               P{GMR10G03-GAL4} expression pattern VFBexp_FBtp0057363
#> 22               P{GMR48H11-GAL4} expression pattern VFBexp_FBtp0060750
#> 23               P{GMR58E02-GAL4} expression pattern VFBexp_FBtp0061564
#> 24                        dopaminergic PAM neuron 12      FBbt_00100230
#> 25                         dopaminergic PAM neuron 3      FBbt_00100231
#> 26                           dopaminergic PAM neuron      FBbt_00100214
#> 27                    PAM-y5(uc)#5_R (FAFB:11149656)       VFB_001014mo
#> 28                    PAM-y5(uc)#4_R (FAFB:11274761)       VFB_001014mt
#> 29                    PAM-y5(da)#1_R (FAFB:11169413)       VFB_001014mp
#> 30                    PAM-y5(uc)#6_R (FAFB:11175371)       VFB_001014mq
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                       synonym
#> 1  peptidylglycine 2-hydroxylase activity, peptidylglycine alpha-hydroxylase activity, peptidylglycine monooxygenase activity, peptide-alpha-amide synthetase activity, PAM-B, peptidylglycine alpha-amidating monooxygenase activity, PAM-A, peptidyl alpha-amidating enzyme activity, peptide alpha-amidating enzyme, PAM activity, peptide alpha-amide synthase activity, synthase, peptide alpha-amide, peptidylglycine,ascorbate:oxygen oxidoreductase (2-hydroxylating)
#> 2                                                                                                                                                                                                                                                                                                 PAM-gamma5 upper commissure neuron, PAM-gamma5(uc), gamma5-uc, PAM-gamma5(da), PAM01 neuron of upper MB commissure, PAM01-uc, dopaminergic PAM neuron 1 of upper commissure
#> 3                                                                                                                                                                     PAM-gamma5(ba), PAM01 neuron of middle commissure, PAM01 neuron of ventral MB commissure, PAM01 neuron of lower commissure, PAM-gamma5 lower commissure neuron, PAM01-lc, PAM-gamma5 middle commissure neuron, gamma5-lc, dopaminergic PAM neuron 1 of ventral mushroom body commissure, PAM-gamma5(lc)
#> 4                                                                                                                                                                                                                                                                                                                                PAM01 feedback neuron, PAM01-fb, PAM-gamma5 dorsal dendrite neuron, PAM-gamma5(fb), PAM-gamma5(dd), PAM-gamma5 feedback neuron, gamma5-lc-fb
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                PAM15, dopaminergic PAM neuron 15, PAM-15, PAM-gamma5beta'2a
#> 6                                                                                                                                                                                                                                                                                                                                larval dopaminergic pPAM neuron, larval PAM neuron, larval DAN of pPAM cluster, primary-lineage protocerebral anterior medial cluster neuron
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                      PAM-beta1ped, PAM09, PAM-09, dopaminergic PAM neuron 9
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                        dopaminergic PAM neuron 8, PAM08, PAM-08, PAM-gamma4
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                       PAM-beta'2m, PAM06, PAM-06, dopaminergic PAM neuron 6
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                      dopaminergic PAM neuron 5, PAM-beta'2p, PAM-05, PAM05
#> 11                                                                                                                                                                                                                                                                                                                                                                                                         beta'2a DAN, dopaminergic PAM neuron 2, PAM-beta'2a, PAM02, PAM-02
#> 12                                                                                                                                                                                                                                                                                                                                                                              PAM-01, adult fruitless aSP13 neuron, PAM01, DAN-aSP13, dopaminergic PAM neuron 1, PAM-gamma5
#> 13                                                                                                                                                                                                                                                                                                                                                                                                          PAM-07, PAM-gamma4-gamma1gamma2, PAM07, dopaminergic PAM neuron 7
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                     PAM-14, dopaminergic PAM neuron 14, PAM-beta'1m, PAM14
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                    PAM-beta'1ap, PAM13, dopaminergic PAM neuron 13, PAM-13
#> 16                                                                                                                                                                                                                                                                                                                                                                                                              dopaminergic PAM neuron 11, MB-VP1, PAM11, PAM-11, PAM-alpha1
#> 17                                                                                                                                                                                                                                                                                                                                                                                                                        dopaminergic PAM neuron 4, PAM-04, PAM-beta2, PAM04
#> 18                                                                                                                                                                                                                                                                                                                                                                                                                       dopaminergic PAM neuron 10, PAM-10, PAM-beta1, PAM10
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                     PRESEQUENCE TRANSLOCASE-ASSOCIATED MOTOR SUBUNITS, PAM
#> 20                                                                                                                                                                                                                                                                                                                                                                                                                             TH negative neuron associated with PAM cluster
#> 21                                                                                                                                                                                                                                                                                                 P{GMR10G03-GAL4} expression pattern, PAM-DAN-γ4-Gal4 expression pattern, R10G03-Gal4 expression pattern, PAM-DAN-gamma4-Gal4 expression pattern, R10G03 expression pattern
#> 22                                                                                                                                                                                                                                                                PAM-DAN-γ5-Gal4 expression pattern, PAM-DAN-gamma5-Gal4 expression pattern, R48H11-Gal4 expression pattern, R48H11 expression pattern, P{GMR48H11-GAL4} expression pattern, R48H011-Gal4 expression pattern
#> 23                                                                                                                                                                          R58E02GAL4 expression pattern, R58E02-Gal4 expression pattern, GMR58E02-Gal4 expression pattern, GMR58E02-GAL4 expression pattern, R58E02 expression pattern, R58E02-GAL4 expression pattern, P{GMR58E02-GAL4} expression pattern, 58E02-GAL4 expression pattern, PAM-DAN-Gal4 expression pattern
#> 24                                                                                                                                                                                                                                                                                                                                                         mushroom body medial lobe arborizing neuron 2, MB-M2 neuron, PAM-12, PAM12, dopaminergic PAM neuron 12, PAM-gamma3
#> 25                                                                                                                                                                                                                                                                                                                                                           MB-M3, PAM-03, PAM-beta2beta'2a, dopaminergic PAM neuron 3, PAM03, mushroom body medial lobe arborizing neuron 3
#> 26                                                                                                                                                                                                                                     protocerebral anteriomedial dopaminergic cluster neuron, PAM, AM cell cluster neuron, dopaminergic PAM neuron, TH-positive anteriomedial cluster neuron, PAM cluster neuron, protocerebral anterior medial dopaminergic cluster neuron
#> 27                                                                                                                                                                                                                                                                                                                                                                                                                                      Putative DAN PAM-y5 (25) 11149657 KJH
#> 28                                                                                                                                                                                                                                                                                                                                                                                                                                      Putative DAN PAM-y5 (24) 11274762 KJH
#> 29                                                                                                                                                                                                                                                                                                                                                                                                                                      Putative DAN PAM-y5 (21) 11169414 KJH
#> 30                                                                                                                                                                                                                                                                                                                                                                                                                                     Putative DAN PAM-y5 (22) 11175372 AJES

# \donttest{
# Search for MBON-01 to MBON-22 (sprintf is used to 0 pad the numbers)
vfb_synonym_query(sprintf("MBON-%02d",1:22))
#> $`MBON-01`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-02`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-03`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-04`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-05`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-06`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-07`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-08`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-09`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-10`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-11`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-12`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-13`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-14`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-15`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-16`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-17`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-18`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-19`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-20`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-21`
#> data frame with 0 columns and 0 rows
#> 
#> $`MBON-22`
#> data frame with 0 columns and 0 rows
#> 

# You can also use a wild card search, which is much faster since it only
# makes a single solr query but the hits are returned in an arbitrary order.
mbondf=vfb_synonym_query("MBON-??")
# then you can pick out your preferred synonym
# note that we use the glob2rx function to convert solr's simple shell-style
# wild card syntax to a regular expression
mbondf$aso=sapply(mbondf$synonym, function(x) grep(glob2rx("MBON-??"), x, value=TRUE))
# }