Convert between VFB and external identifiers

Convert between VFB and external identifiers

vfb_fromvfbids(vfbids, ...)

vfb_tovfbids(ids, fixed = TRUE, ...)

Arguments

vfbids

One or more vfb identifiers. Solr wildcards can also be used when fixed=FALSE- see examples and vfb_solr_query

...

Additional arguments passed to vfb_solr_query.

ids

One or more external identifiers. Solr wildcards can also be used when fixed=FALSE- see examples and vfb_solr_query

fixed

Whether to insist on exact matches. When fixed=FALSE solr wildcards can be used.

Value

A character vector of vfb ids (for vfb_fromvfbids) or external identifiers (for vfb_tovfbids) in the same order as the input query. Missing values will be denoted by NAs when fixed=TRUE. When fixed=FALSE and a wild-card search returns no results a character vector of length 0 will be returned.

Details

Note that these functions are not appropriate for fetching more than ~ 200 identifiers individually specified identifiers. If a large number of identifiers are required, it is a much better idea to fetch all ids and then filter locally.

See also

gmr_vfbid

Examples

# some flycircuit ids
fcids=c("VGlut-F-000304", "VGlut-F-200278", "fru-F-200121", "TH-F-300016")
vfbids=vfb_tovfbids(fcids)
vfb_fromvfbids(vfbids)

#> [1] "VGlut-F-000304" "VGlut-F-200278" "fru-F-200121"   "TH-F-300016"