Query VFB's Neo4J graph database

Query VFB's Neo4J graph database

vfb_neo4j_query(x, ..., path = "neo4jdb/data",
  server = getOption("vfbr.server"))

Arguments

x

A character query in Neo4J's cypher language

...

Additional query arguments of the form key=value

path

The relative path on the server for the Neo4J endpoint

server

The server's root URL

Value

A data.frame of query results

Details

Under the hood, this uses the cypher function to call a Neo4J service running on the specified VFB server. The path and server arguments should be

References

https://neo4j.com/developer/cypher-query-language/

See also

RNeo4j, cypher

Other query: vfb_owl_query, vfb_solr_query

Examples

# ask for all neuronal classes
nclasses=vfb_neo4j_query("MATCH (n:Neuron:Class) RETURN n.label")
nrow(nclasses)

#> [1] 2687

head(nclasses)

#>                               n.label
#> 1               abdominal vch2 neuron
#> 2            abdominal sensory neuron
#> 3 abdominal posterior fascicle neuron
#> 4                         vch2 neuron
#> 5             abdominal 7 vch2 neuron
#> 6                           A7 neuron

# Find all images with an associated neuronal class
q=paste0("MATCH (n:Class:VFB { label : 'neuron' })",
  "<-[:SUBCLASSOF*]-(p)<-[:INSTANCEOF]-(i:Individual)",
  "RETURN distinct i.label, p.label;")
nclasses_image=vfb_neo4j_query(q)