Query VFB's Neo4J graph database

Query VFB's Neo4J graph database

vfb_neo4j_query(x, ..., path = "db/data",
  server = getOption("vfbr.server.neo4j"))

Arguments

x
A character query in Neo4J's cypher language

...
Additional query arguments of the form key=value

path
The relative path on the server for the Neo4J endpoint

server
The server's root URL

Value

A data.frame of query results

Details

Under the hood, this uses the cypher function to call a Neo4J service running on the specified VFB server. The path and server arguments should be

References

https://neo4j.com/developer/cypher-query-language/

See also

Examples

# ask for all neuronal classes
nclasses=vfb_neo4j_query("MATCH (n:Neuron:Class) RETURN n.label")
nrow(nclasses)

#> [1] 2772

head(nclasses)

#>                                                       n.label
#> 1                              mushroom body intrinsic neuron
#> 2 transverse tract multiglomerular projection neuron AL-t2PN1
#> 3       adult multiglomerular antennal lobe projection neuron
#> 4            transverse antennal lobe tract projection neuron
#> 5                              mushroom body extrinsic neuron
#> 6                       adult antennal lobe projection neuron

# Find all images with an associated neuronal class
q=paste0("MATCH (n:Class:VFB { label : 'neuron' })",
  "<-[:SUBCLASSOF*]-(p)<-[:INSTANCEOF]-(i:Individual)",
  "RETURN distinct i.label, p.label;")
nclasses_image=vfb_neo4j_query(q)
nrow(nclasses_image)

#> [1] 7108

head(nclasses_image)

#>              i.label                                                p.label
#> 1 MB-MV2 of Aso 2014 mushroom body medial-vertical lobe arborizing neuron 2
#> 2  MB-M4 of Aso 2014          mushroom body medial lobe arborizing neuron 4
#> 3  MB-M6 of Aso 2014          mushroom body medial lobe arborizing neuron 6
#> 4     VGlut-F-500350                            lobula columnar neuron Lcn4
#> 5     VGlut-F-200028                            lobula columnar neuron Lcn4
#> 6       Cha-F-700192                            lobula columnar neuron Lcn4

# how many neuronal classes have images?
length(unique(nclasses_image$p.label))

#> [1] 280

# look at number of images for each neuronal class
table_by_nclass <- table(nclasses_image$p.label)
subset(as.data.frame(table_by_nclass), Freq>200)

#>                                                          Var1 Freq
#> 36                                   adult DM6 lineage neuron  295
#> 74                                  adult fruitless Mb neuron  682
#> 102 adult multiglomerular antennal lobe projection neuron lPN  201
#> 120                               alpha/beta core Kenyon cell  689
#> 122                            alpha/beta surface Kenyon cell  366
#> 192                                         gamma Kenyon cell  423
#> 199                               lobula columnar neuron LC12  212