Query VFB via solr indexing system

Solr provides an extremely fast way to query all key content on VFB and is the backend used for most queries run on the website. It is pre-populated from the OWL ontology documents describing all the information on the site and their relationships. More details on Solr and its query syntax can be found at http://lucene.apache.org/solr/.

vfb_solr_query(query = "*:*", filterquery = NULL,
  fields = "label+short_form", sort = "score+desc",
  defaultfield = "short_form", rows = 30L,
  path = "solr/ontology/select?wt=json",
  server = getOption("vfbr.server.solr"), parse.json = TRUE, ...)

Arguments

query
A key-value list specifying the query

filterquery
A character vector (of length one or more) describing filter queries for solr (see Details for regular vs filter queries)

fields
Which fields to return (+delimited). A value of "" implies all fields.

sort
Character vector naming one or more fields (+ delimited) to use for sorting the results.

defaultfield
Character vector naming default field used for filter queries (defaults to short_form)

rows
Maximum number of rows to return. The special value of Inf implies all matching rows.

path
The path on the server containing the query page

server
The base url of the server

parse.json
Whether or no to parse the response (default: TRUE)

...
additional solr query arguments

Value

When parse.json=TRUE, a data.frame containing the parsed response (originally the response$docs field in the parsed JSON) along with additional attributes including

numFound

start

responseHeader

When parse.json=FALSE an httr::response object

Details

The query arguments maps onto the general solr q= query while filterqueries maps onto one or more fl= terms. The solr wiki says this about the difference:

The fq parameter defines a query that can be used to restrict the superset of documents that can be returned, without influencing score. It can be very useful for speeding up complex queries, since the queries specified with fq are cached independently of the main query. When a later query uses the same filter, there's a cache hit, and filter results are returned quickly from the cache.

See also

Examples

# Find VFB ids matching a given GMR line
# note the field synonym_autosuggest will in future be the only one
# matching GMR* ids
vfb_solr_query(filterquery="VFB_*",query="synonym_autosuggest:GMR_10A07*")

#>     short_form                    label
#> 1 VFB_00023198 GMR_10A07_AE_01_09-fA01b

# Find VFB ids matching a given VT Gal4 line
vfb_solr_query(filterquery="VFB_*",query="label:VT017929*")

#>     short_form                                        label
#> 1 VFB_00064845   VT017929 GAL4 in the Adult VNC Dickson2017
#> 2 VFB_00064844 VT017929 GAL4 in the Adult Brain Dickson2017

# how many GMR lines can we find
# note use of rows = 0 so we do not fetch results (but still get totals)
r=vfb_solr_query(filterquery="VFB_*",query="label:GMR_*", rows=0)
attr(r,'numFound')

#> [1] 3468

#' # VFB id for all GMR lines
all_gmr=vfb_solr_query(filterquery="VFB_*",query="label:GMR_*", rows=4000)
head(all_gmr)

#>     short_form                    label
#> 1 VFB_00021308 GMR_16B10_AE_01_05-fA01b
#> 2 VFB_00021309 GMR_85A07_AE_01_02-fA01b
#> 3 VFB_00021306 GMR_12H02_AE_01_07-fA01b
#> 4 VFB_00021307 GMR_42D11_AE_01_05-fA01b
#> 5 VFB_00021311 GMR_41F05_AE_01_05-fA01b
#> 6 VFB_00021312 GMR_36E07_AE_01_00-fA01b

# VFB id for all FlyCircuit neurons
# note use of rows=Inf to fetch all rows
all_fc=vfb_solr_query(filterquery="VFB_*",
  query="source_data_link_annotation:*flycircuit*", rows=Inf)
head(all_fc)

#> data frame with 0 columns and 0 rows