Make a URL for a set of VFB ids to be displayed in the stack browser

Make a URL for a set of VFB ids to be displayed in the stack browser

vfb_stack_url(ids, terminfo = NULL, clear = FALSE,
  path = "/site/stacks", server = getOption("vfbr.server"))

Arguments

ids
A character vector of IDs

terminfo
A VFB id used to choose which term information will be displayed next to the stack browser.

clear
Whether to clear items already displayed in the stack browser (default FALSE)

path
The path on the server containing the query page

server
The base url of the server

Value

character vector to open stack browser

Examples

# some gmr lines
ids=c("VFB_00004657","VFB_00023207","VFB_00023120","VFB_00022264")
template="VFBt_00100000"
u=vfb_stack_url(c(template, ids))

# find vfb id for this class of neuron often called fruitless mAL neuron
resdf=vfb_solr_query("synonym:mAL",filterquery = c("label:female","label:neuron"))
head(resdf)

#>      short_form                                 label
#> 1 FBbt_00110417 adult fruitless aDT-b (female) neuron

u=vfb_stack_url(c(template, ids), terminfo=resdf$short_form[1], clear=TRUE)
# NOT RUN {
browseURL(u)
# }