Find canononical ontology term using VFB SOLR to query synonyms

Find canononical ontology term using VFB SOLR to query synonyms

vfb_synonym_query(query, exact = TRUE, quote = NA,
  searchfields = c("synonym", "label"), fields = "short_form label synonym",
  verbose = interactive(), ...)

Arguments

query
A character vector containing one or more queries - see details. Wildcards (?*) can be used, and have a similar meaning to their use in the shell. Queries are case insensitive.

exact
Whether to do an exact match (i.e. the query must match the whole string, the default) or to allow partial matches against synonyms. See Query details and examples.

quote
Whether to quote the query so that individual terms must all be matched in the order given or to OR each term in the query. The default quote=NA will quote when exact=TRUE and the query does not contain wildcards. See Query details and Examples.

searchfields
Character vector specifying fields to search. The default searches both synonyms and the canonical term label (since the canonical term is not included in the synonym list).

fields
The fields to return (defaults to 'short_form label synonym'). Setting fields="" implies all fields (but note that not all results may have the same fields - see discussion in details)

verbose
Whether to print messages to the screen.

...
Additional arguments passed to vfb_solr_query. You can use this e.g. to set the number of returned rows.

Value

A data.frame containing one or more result rows, ordered according to the solr result score, with attributres

Details

Note that data.frame that is returned may contain a list in the synonym column because there will likely be multiple synonyms for a given query.

When x has length > 1 i.e. multiple query terms, then multiple calls to vfb_solr_query are wrapped in a sapply statement. You can pass arguments to sapply in ... such as simplify=FALSE if you wish. By default the return value will be a matrix with rows that you can index by field name and columns that you can index by queries. However the form of this is not guaranteed if you ask for fields that are present only for some of the results.

Query details

The exact and quote arguments have a profound effect on how SOLR carries out searches. Setting both arguments to FALSE gives the least specific search and you may want to do this if you do not find any results with an initial query using the default values.

In general SOLR will break any text in fields like synonym into separate tokens based on whitespace, hypens etc, so the value :

inter-antennal lobe tract

will map onto 4 tokens.

inter antennal lobe tract

This means that a search with query "inter antennal lobe tract", exact=F will be successful (note missing hyphen).

When exact=TRUE, the values will not be tokenised and the query must match the whole field.

The quote argument modifies the query. By default the query is tokenised as just described for fields in the database, tokens can be matched in any order and only one token must match (i.e. the are ORed together). When quote=TRUE, the query must match in the order given (this is achieved by wrapping it in double quotes before passing to SOLR). For example

vfb_synonym_query("antennal lobe tract", exact=F, quote=T)

returns 2 results at the time of writing whereas

vfb_synonym_query("antennal lobe tract", exact=F, quote=F)

returns 194 results.

See https://github.com/EBISPOT/OLS/blob/master/ols-solr/src/main/solr-conf/ontology/conf/schema.xml for definition of the of the synonym_s exact match field.

See also

Examples

vfb_synonym_query("SOG")

#>      short_form                    label
#> 1 FBbt_00110639 adult subesophageal zone
#> 2 FBbt_00004013   subesophageal ganglion
#>                                                          synonym
#> 1 suboesophageal ganglion, subesophageal ganglion, SEG, SEZ, SOG
#> 2                              suboesophageal ganglion, SEG, SOG

# anything with a synonym that includes SOG
vfb_synonym_query("SOG", exact=FALSE)

#>      short_form                                   label
#> 1 FBbt_00110155              octopaminergic VUMa neuron
#> 2 FBbt_00111345   larval CCAP subesophageal IN-1 neuron
#> 3 FBbt_00111326        larval CCAP subesophageal neuron
#> 4 FBbt_00110639                adult subesophageal zone
#> 5 FBbt_00100785      SOG ventral unpaired median neuron
#> 6 FBbt_00100138 embryonic/larval subesophageal ganglion
#> 7 FBbt_00004013                  subesophageal ganglion
#> 8 FBbt_00007354             antenno-subesophageal tract
#>                                                                                                                  synonym
#> 1                                                             OA-VUMa, subesophageal ganglion VUM neuron, SOG VUM neuron
#> 2                                                                                                               IN SOG-1
#> 3                                                                                                                 IN SOG
#> 4                                                         suboesophageal ganglion, subesophageal ganglion, SEG, SEZ, SOG
#> 5                                                                      subesophageal ganglion VUM neuron, SOG VUM neuron
#> 6                                                                                                   embryonic/larval SOG
#> 7                                                                                      suboesophageal ganglion, SEG, SOG
#> 8 AST, antennal lobe-subesophageal tract, antennal lobe-SOG tract, antenno-suboesophageal tract, antennal lobe-SEG tract

# query for PAM exactly
vfb_synonym_query("PAM")

#>      short_form                   label
#> 1 FBbt_00100214 dopaminergic PAM neuron
#>                                                                                                                                              synonym
#> 1 PAM cluster neuron, AM cell cluster neuron, TH-positive anteriomedial cluster neuron, PAM, protocerebral anteriomedial dopaminergic cluster neuron

# any term with a synonym containing PAM
vfb_synonym_query("PAM", exact = FALSE)

#>       short_form                                         label
#> 1  FBbt_00111024                    dopaminergic PAM neuron 14
#> 2  FBbt_00111023                    dopaminergic PAM neuron 13
#> 3  FBbt_00111022                    dopaminergic PAM neuron 11
#> 4  FBbt_00111018                     dopaminergic PAM neuron 6
#> 5  FBbt_00111017                     dopaminergic PAM neuron 5
#> 6  FBbt_00111016                     dopaminergic PAM neuron 2
#> 7  FBbt_00111015                     dopaminergic PAM neuron 1
#> 8  FBbt_00111019                     dopaminergic PAM neuron 7
#> 9  FBbt_00111021                     dopaminergic PAM neuron 9
#> 10 FBbt_00111020                     dopaminergic PAM neuron 8
#> 11 FBbt_00110993                     dopaminergic PAM neuron 4
#> 12 FBbt_00110994                    dopaminergic PAM neuron 10
#> 13 FBbt_00100230 mushroom body medial lobe arborizing neuron 2
#> 14 FBbt_00100214                       dopaminergic PAM neuron
#> 15 FBbt_00100231 mushroom body medial lobe arborizing neuron 3
#>                                                                                                                                               synonym
#> 1                                                                                                                                 PAM-beta'1m, PAM-14
#> 2                                                                                                                                PAM-13, PAM-beta'1ap
#> 3                                                                                                                                  PAM-alpha1, PAM-11
#> 4                                                                                                                                 PAM-beta'2m, PAM-06
#> 5                                                                                                                                 PAM-beta'2p, PAM-05
#> 6                                                                                                                                 PAM-beta'2a, PAM-02
#> 7                                                                                                                                  PAM-01, PAM-gamma5
#> 8                                                                                                                     PAM-07, PAM-gamma4-gamma1gamma2
#> 9                                                                                                                                PAM-09, PAM-beta1ped
#> 10                                                                                                                                 PAM-08, PAM-gamma4
#> 11                                                                                                                                  PAM-beta2, PAM-04
#> 12                                                                                                                                  PAM-10, PAM-beta1
#> 13                                                                                                                          MB-M2, PAM-12, PAM-gamma3
#> 14 PAM cluster neuron, AM cell cluster neuron, TH-positive anteriomedial cluster neuron, PAM, protocerebral anteriomedial dopaminergic cluster neuron
#> 15                                                                                                                    MB-M3, PAM-beta2beta'2a, PAM-03

# terms with synonyms that start with "PAM-"
vfb_synonym_query("PAM-*")

#>       short_form                                         label
#> 1  FBbt_00111024                    dopaminergic PAM neuron 14
#> 2  FBbt_00111023                    dopaminergic PAM neuron 13
#> 3  FBbt_00111022                    dopaminergic PAM neuron 11
#> 4  FBbt_00111018                     dopaminergic PAM neuron 6
#> 5  FBbt_00111017                     dopaminergic PAM neuron 5
#> 6  FBbt_00111016                     dopaminergic PAM neuron 2
#> 7  FBbt_00111015                     dopaminergic PAM neuron 1
#> 8  FBbt_00111019                     dopaminergic PAM neuron 7
#> 9  FBbt_00111021                     dopaminergic PAM neuron 9
#> 10 FBbt_00111020                     dopaminergic PAM neuron 8
#> 11 FBbt_00110993                     dopaminergic PAM neuron 4
#> 12 FBbt_00110994                    dopaminergic PAM neuron 10
#> 13 FBbt_00100230 mushroom body medial lobe arborizing neuron 2
#> 14 FBbt_00100231 mushroom body medial lobe arborizing neuron 3
#>                            synonym
#> 1              PAM-beta'1m, PAM-14
#> 2             PAM-13, PAM-beta'1ap
#> 3               PAM-alpha1, PAM-11
#> 4              PAM-beta'2m, PAM-06
#> 5              PAM-beta'2p, PAM-05
#> 6              PAM-beta'2a, PAM-02
#> 7               PAM-01, PAM-gamma5
#> 8  PAM-07, PAM-gamma4-gamma1gamma2
#> 9             PAM-09, PAM-beta1ped
#> 10              PAM-08, PAM-gamma4
#> 11               PAM-beta2, PAM-04
#> 12               PAM-10, PAM-beta1
#> 13       MB-M2, PAM-12, PAM-gamma3
#> 14 MB-M3, PAM-beta2beta'2a, PAM-03

# nb this doesn't return any results when exact=FALSE because solr matches
# against tokenised strings (i.e. strings that have broken on spaces and
# other non word characters such as dashes and underscores).
vfb_synonym_query("PAM-*", exact = FALSE)

#> data frame with 0 columns and 0 rows

# However if you quote the query you will get results
vfb_synonym_query("PAM-*", exact = FALSE, quote=TRUE)

#>       short_form                                         label
#> 1  FBbt_00111024                    dopaminergic PAM neuron 14
#> 2  FBbt_00111023                    dopaminergic PAM neuron 13
#> 3  FBbt_00111022                    dopaminergic PAM neuron 11
#> 4  FBbt_00111018                     dopaminergic PAM neuron 6
#> 5  FBbt_00111017                     dopaminergic PAM neuron 5
#> 6  FBbt_00111016                     dopaminergic PAM neuron 2
#> 7  FBbt_00111015                     dopaminergic PAM neuron 1
#> 8  FBbt_00111019                     dopaminergic PAM neuron 7
#> 9  FBbt_00111021                     dopaminergic PAM neuron 9
#> 10 FBbt_00111020                     dopaminergic PAM neuron 8
#> 11 FBbt_00110993                     dopaminergic PAM neuron 4
#> 12 FBbt_00110994                    dopaminergic PAM neuron 10
#> 13 FBbt_00100230 mushroom body medial lobe arborizing neuron 2
#> 14 FBbt_00100214                       dopaminergic PAM neuron
#> 15 FBbt_00100231 mushroom body medial lobe arborizing neuron 3
#>                                                                                                                                               synonym
#> 1                                                                                                                                 PAM-beta'1m, PAM-14
#> 2                                                                                                                                PAM-13, PAM-beta'1ap
#> 3                                                                                                                                  PAM-alpha1, PAM-11
#> 4                                                                                                                                 PAM-beta'2m, PAM-06
#> 5                                                                                                                                 PAM-beta'2p, PAM-05
#> 6                                                                                                                                 PAM-beta'2a, PAM-02
#> 7                                                                                                                                  PAM-01, PAM-gamma5
#> 8                                                                                                                     PAM-07, PAM-gamma4-gamma1gamma2
#> 9                                                                                                                                PAM-09, PAM-beta1ped
#> 10                                                                                                                                 PAM-08, PAM-gamma4
#> 11                                                                                                                                  PAM-beta2, PAM-04
#> 12                                                                                                                                  PAM-10, PAM-beta1
#> 13                                                                                                                          MB-M2, PAM-12, PAM-gamma3
#> 14 PAM cluster neuron, AM cell cluster neuron, TH-positive anteriomedial cluster neuron, PAM, protocerebral anteriomedial dopaminergic cluster neuron
#> 15                                                                                                                    MB-M3, PAM-beta2beta'2a, PAM-03

# Search for MBON-01 to MBON-22 (sprintf is used to 0 pad the numbers)
vfb_synonym_query(sprintf("MBON-%02d",1:22))

#>            MBON-01                                        
#> short_form "FBbt_00100234"                                
#> label      "mushroom body medial lobe arborizing neuron 6"
#> synonym    List,1                                         
#>            MBON-02                        
#> short_form "FBbt_00111012"                
#> label      "mushroom body output neuron 2"
#> synonym    List,1                         
#>            MBON-03                                        
#> short_form "FBbt_00100232"                                
#> label      "mushroom body medial lobe arborizing neuron 4"
#> synonym    List,1                                         
#>            MBON-04                         MBON-05                        
#> short_form "FBbt_00111014"                 "FBbt_00111004"                
#> label      "mushroom body output neuron 4" "mushroom body output neuron 5"
#> synonym    List,1                          List,1                         
#>            MBON-06                                                 
#> short_form "FBbt_00100242"                                         
#> label      "mushroom body medial-vertical lobe arborizing neuron 2"
#> synonym    List,1                                                  
#>            MBON-07                         MBON-08                        
#> short_form "FBbt_00111005"                 "FBbt_00111006"                
#> label      "mushroom body output neuron 7" "mushroom body output neuron 8"
#> synonym    List,1                          List,1                         
#>            MBON-09                         MBON-10                         
#> short_form "FBbt_00111007"                 "FBbt_00111008"                 
#> label      "mushroom body output neuron 9" "mushroom body output neuron 10"
#> synonym    List,1                          List,1                          
#>            MBON-11                                                                     
#> short_form "FBbt_00100246"                                                             
#> label      "mushroom body pedunculus-medial lobe and vertical lobe arborizing neuron 2"
#> synonym    List,1                                                                      
#>            MBON-12                         
#> short_form "FBbt_00111009"                 
#> label      "mushroom body output neuron 12"
#> synonym    List,1                          
#>            MBON-13                                          
#> short_form "FBbt_00100239"                                  
#> label      "mushroom body vertical lobe arborizing neuron 4"
#> synonym    List,1                                           
#>            MBON-14                                          
#> short_form "FBbt_00100238"                                  
#> label      "mushroom body vertical lobe arborizing neuron 3"
#> synonym    List,1                                           
#>            MBON-15                         
#> short_form "FBbt_00111010"                 
#> label      "mushroom body output neuron 15"
#> synonym    List,1                          
#>            MBON-16                                                                    
#> short_form "FBbt_00111063"                                                            
#> label      "mushroom body vertical lobe arborizing neuron 2 alpha' anterior-posterior"
#> synonym    List,1                                                                     
#>            MBON-17                                                        
#> short_form "FBbt_00111064"                                                
#> label      "mushroom body vertical lobe arborizing neuron 2 alpha' middle"
#> synonym    List,1                                                         
#>            MBON-18                                                
#> short_form "FBbt_00110101"                                        
#> label      "mushroom body vertical lobe arborizing neuron 2 alpha"
#> synonym    List,1                                                 
#>            MBON-19                          MBON-20                         
#> short_form "FBbt_00111011"                  "FBbt_00111013"                 
#> label      "mushroom body output neuron 19" "mushroom body output neuron 20"
#> synonym    List,1                           List,1                          
#>            MBON-21                         
#> short_form "FBbt_00111046"                 
#> label      "mushroom body output neuron 21"
#> synonym    List,1                          
#>            MBON-22                                             
#> short_form "FBbt_00100240"                                     
#> label      "mushroom body calyx-pedunculus arborizing neuron 1"
#> synonym    List,1                                              

# You can also use a wild card search, which is much faster since it only
# makes a single solr query but the hits are returned in an arbitrary order.
mbondf=vfb_synonym_query("MBON-??")
# then you can pick out your preferred synonym
# note that we use the glob2rx function to convert solr's simple shell-style
# wild card syntax to a regular expression
mbondf$aso=sapply(mbondf$synonym, function(x) grep(glob2rx("MBON-??"), x, value=TRUE))