Programmatic Access to the virtualflybrain.org website

Programmatic Access to the virtualflybrain.org website

Queries

vfb_solr_query Allows you to query VFB's SOLR end point. This gives you programmatic access to the great majority of searches that you can perform interactively on the VFB website.

vfb_neo4j_query Allows you to query VFB's Neo4J end point. This gives you programmatic access to many sophisticated searches based on the VFB ontology, including searches that may not be easily accessible via the website.

vfb_synonym_query A higher level function that allows you to search for canonical terms with synonyms matching a given query.

Package Options

The following options can be set to specify default behaviour. They will be set to sensible defaults on package startup, unless you have previously set a value (e.g. in your Rprofile).

vfbr.server URL of main VFB server

vfbr.server.gepetto URL of VFB's gepetto server for 3D visualisation (see vfb_3dbrowser_url.)

vfbr.stack.gmr_urlURL containing listing of registered GMR Gal4 confocal stacks. See gmr_stack_urls

vfbr.stack.downloadsLocation of downloaded stacks. See download_gmr_stacks

References

See http://virtualflybrain.org

See also

vfb_solr_query

Examples

# Show state of vfbr package options
options()[grep('^vfbr', names(options()))]

#> $vfbr.server
#> [1] "http://vfbdev.inf.ed.ac.uk"
#> 
#> $vfbr.server.gepetto
#> [1] "http://129.215.164.244:8084"
#> 
#> $vfbr.stack.downloads
#> [1] "/Users/jefferis/Library/Application Support/rpkg-vfbr/stacks"
#> 
#> $vfbr.stack.gmr_url
#> [1] "http://flybrain.mrc-lmb.cam.ac.uk/vfb/jfrc/fl/reformatted-quant/"
#> 

## Not run: ------------------------------------
# example(vfb_solr_query)
## ---------------------------------------------